// src/app/events/models/evenement.model.ts
export type Evenement = {
  id?: number | null;
  title: string;
  description: string;
  startDate: Date;
  location: string;

  /*attachment?: {
    name: string;
    content: string; // base64
    type: string;
  }*/

  /*
  endDate: Date;
  imageUrl?: string;
  isActive: boolean;
  isPublic: boolean;
  createdAt: Date;
  updatedAt: Date;
 */
};